Updated to allow -- in data dir name
Right now, gpstart relies on "--" as a STATUS field splitter when
interpreting the results of starting the segments.  The issue is that
STATUS field(in the SegStopStatus as  `STATUS--DIR:%s`) contains a
directory name.  But a directory name can contain `--`.

The current parsing code does not handle this well.
We have improved the parsing code to allow for `--` in dir names