Convert policy of root to random when one of its children has a different distribution policy. (#12671)
The only allowed difference between parent and children is for the
parent to be hash distributed, and its child part to be randomly distributed.
When the policy is different between parent and children, The policy of
root is converted to random in GPDB6. However, relevant codes are
commented out in MASTER. We can not alter the policy for child table.
But when we support expand partition table in parallel, The policy of
child and parent may be different, Parent is hash distributed and child
is random. In this situation, We should convert the policy of root to
random when the policy between parent and children is different to avoid
wrong results.

This works in GDPB6, master branch misses this during Postgres v12 merge.
This commit adds back this. For details, please refer the mailing list thread:
https://groups.google.com/a/greenplum.org/g/gpdb-dev/c/rSacd_vI-fM/m/pkAW-Z-lCgAJ