ORCA: Avoid divide-by-zero when intersecting histogram buckets (#12751)
When intersecting two empty histogram buckets, we expect the result
histogram bucket to be empty too.

This patch avoids the edge case of CDouble(0) / CDouble(0) = 1 when
calculating the m_frequency of the result bucket.

This patch also expands coverage of the CBucketIntersetTest unit test. 
Previously we only check the correctness of bucket boundaries in the
unit test for CBucket::MakeBucketIntersect(), this commit verifies the
correctness of frequencies and NDVs as well for the existing test cases,
and adds new test case for intersecting empty histogram buckets.

Fixes #12750.