Fix multiple port issue in gpexpand also added health check warning
Issue: when one or more cluster is down or not in their preferred role,
the assignment of port on expanded nodes are going wrong. the reason behind
that the calculation of port using function get_max_primary_port() and
get_max_mirror_port(). The idea to exit from gpexpand when the cluster is
not in balanced state was blocking customer to expand if they found difficulty
to make the system in balance state.

Fix: updated the function to work in imbalance cluster state scenario.
we are collecting all the port for primary and mirror based on preferred role
of the segment, and if the role is switched we are considering segmentPairs.mirrorDB
for primary port and segmentPairs.primaryDB for mirror port.
in this we are collecting the port which is based on the preferred role.

Added warning in gpexpand when the cluster is in imbalance state.

Added unit test cases for the positive/negative and exception scenario for the newly
added functions. Updated feature test case where we are converting error to warning.