Updated slot creation in gprecoverseg full recovery.
Issue: If the cluster never has mirrors, PgBasebackUp call will fail with create-slot option False.
To add the mirror as part of fall back we were making the PgBasebackUp call with create-slot option True.
Fixme was added to fix this code structure.

Fix: We did few spike to check if the slot is present and create-slot is true can it reuse the existing slot.
but there is no support for that. As a optimal solution we are now
droping the available existing slot and creating new if the create-slot
option is provided as True. Which will also help to clean up the orphaned slot.

Implementation details:
Added class PgReplicationSlot which has utility functions to check
if the slot exists and dropping the slot.

Added unit test cases for the class PgReplicationSlot covering all
required code path.

Added test case to TestUnitPgBaseBackup class for the check of PgBaseBackup
command

Added step to verify if the segment has replication slot and included in
behave test of gpmovemirror, gpaddmirror and gprecoverseg

Added scenario to verify the drop slot behaviour.