Added -F to run full recovery in place of incremental
Reason: In this feature test case if the full recovery fails we were trying to run
the pgrewind, with the slot recreate changes we are dropping the replication
slot, as the slot already drop pg_rewing throwing error.
this can be happen  even without the slot recreate code changes,
If in between of full recovery the slot gets deleted or move or renamed.
the possible future solution could be having slot name such that it indicates
mapping of primary and mirror slot, based on that we can use the slot in pg rewind.
but that's the little complex to work on now.
for now in this situation customer would try to fix the issue and rerun
the full recovery instead of incremental.