gpcheckperf: Added support for rsync
Problem : gpcheckperf fails if scp is not available or does
not have execute permission on the host. The reason being it
uses scp to copy over files to remote host to run performance test.

Solution : rsync can be used to copy files to remote host if scp
is not available or does not have execute permisssion.As rsync is
fast and extraordinarily versatile file copying tool which can copy
locally and to remote host as well.

Implementation :
 * check if scp is available and has execute permission on all the hosts
   involved in gpcheckperf operation including master

 * 'which scp' can be used to determine availability of scp as :
    * it will return code 0 if scp is available and has execute permission
    * and will return code 1 if scp is not available or does not have execute permission

 * based on the return code gpcheckperf can decide whether to run scp or rsync:
    * if return code is 0 means scp available so gpcheckperf will execute gpscp()
    * if return code is 1  means scp not available so gpcheckperf will execute gpsync()

Used rsync options:
-a: archive mode
-c : copy files based on checksum
-e: to specify ssh options
-q : quiet mode

Added unit and behave tests for above scenario