Remove gp_toolkit_test.sql
It seems that this file is used nowhere. Many of the views are outdated.
We have a regress/gp_toolkit test which covers most of what
gp_toolkit_test.sql is supposed to test. So removing gp_toolkit_test.sql,
and adding those exist there but not regress/gp_toolkit to regress/gp_toolkit.