Update flakey av ao aux test.
Waiting for fault blocks stats updates sent from the same connection.  There was a race condition in
the test where if we got to the wait before the stats were updated they'd never be received and so
autovacuum would never run and the test would fail on timeout.

This was only visible in greenplum configurations that are faster, most debug builds never observed
this.  The JIT pipeline reliably reproduced it, however.