Skip full recovery to new host if pg_rewind already running
Segments for which pg_rewind is already executing will be skipped during
in-place full and incremental recovery. However, if performing full recovery
to a new host (failover), it will not be skipped.

We should also skip full recovery to a new host/data_dir if pg_rewind is
already running because the check for existing pg_basebackup/pg_rewind
process is source bound (the existence of these processes can be
checked by querying the primary segment) and these query results
provide no information about the target.