Add gp_quicklz_fallback GUC and tests
quicklz compression is not supported in GPDB7.
If the GUC gp_quicklz_fallback = false, setting compresstype=quicklz via CREATE/ALTER/gp_default_storage_options will result in an ERROR.
Setting gp_quicklz_fallback = true will cause the server to fall back to using compresstype=zstd rather than reporting an ERROR.

gp_quicklz_fallback can be enabled to ensure compatibility of pre-GPDB7 backups that may include quicklz compression DDL.

If the server is configured using (--with-zstd) we expect to
fall back to compresstype=zstd when gp_quicklz_fallback=true.

If not, and it's configured using (--with-zlib), fall back
to compresstype=zlib.

If neither are configured, fall back to compresstype=non.

The regression test exercises the fall back logic, but does not
care which compresstype we use, so long as it's valid.

Authored-by: Brent Doil <bdoil@vmware.com>