Improve run time performance for gprecoverseg differential recovery.
Issue: current differential recovery uses rsyn checksum(-c) option in
rsync. this option was added to avoid any data loss as the rsync
checksum scan for diff even if the size and timestamp is same. so -c
option was used to avoid the scenario when for a file the size and time
stamp is same in src and target but content is different.
This scenario can happen when some query is running on primary and the
coordinator connection to primary is lost but still the query is
running. at the same time the mirror got promoted to primary and there
is some other query is running. so in this scenario there is possibility
that the file on primary and mirror can be updated at the same time with
same size content but the content is different. in this case without -c
optiont the rsync will ignore the file to be transferred from source to
target and can cause data loss scenario.

Fix: to avoid above scenario of data loss and improve the performance at
the same time. there is approach in which instead of syncing all the files
using checksum we can do the checksum for only target files, those
target files can be the files which got updated after primary got down.
other files can be sync without checksum option.
for the same added new function sync_updated_files() .

we were also syncing pg_wal twice that can be done at the end.
there is no point in syncing pg_control file as the file will have wrong
information because it will have primary related pg_control info.
also we are reading checkpoint from backup.label file for wal replay
point.

Updated existing test case for the above related changes.