Fix: utilities does not honor -d flag when COORDINATOR_DATA_DIRECTORY is not set.
Issue: following are the gpMgmt utilities does not honor -d flag, when
COORDINATOR_DATA_DIRECTORY is not set.
1. gpstart
2. gpstop
3. gpstatus
4. gprecoverseg
5. gpaddmirror

RCA: to get the coordinator data directory gp.getcoordinator_datadir()
function is called from the above listed utilties. the function do not
have any provision to return the coordinator data directory which is
provided with -d flag. currently it looks for COORDINATOR_DATA_DIRECTORY
and MASTER_DATA_DIRECTORY env variable.

    also in some of the utilities we were creating lock file before
parsing the provided options which looks like design flow which was
causing the utilities to be crash when look for coordinator data
directory.

Fix: added a global flag which holds the datadirectory provided with -d
option. so when we run utility and do parsing it set the flag with
provided datadirectory and the same will be return when we call
gp.gp.getcoordinator_datadir().

Test: Added behave test cases to support the code changes.