Fix utilities do not honor -d flag when MASTER_DATA_DIRECTORY is not set. (#16433)
Issue: Following are the gpMgmt utilities that do not honor the -d flag, when
MASTER_DATA_DIRECTORY is not set.
1. gpstart
2. gpstop
3. gpstatus
4. gprecoverseg
5. gpaddmirror

RCA: to get the master data directory gp.getmaster_datadir()
function is called from the above-listed utilities. the function does not
have any provision to return the master data directory which is
provided with the -d flag. currently, it looks for MASTER_DATA_DIRECTORY
and MASTER_DATA_DIRECTORY env variable.

    also in some of the utilities we were creating lock files before
parsing the provided options which looks like the design flow that was
causing the utilities to crash when looking for master data
directory.

Fix: Added a global flag which holds the data directory provided with -d
option. so when we run the utility and do parsing it sets the flag with
the provided datadirectory and the same will be returned when we call
gp.gp.getmaster_datadir().

Test:
Added behave test cases to use the provided data directory if MASTER_DATA_DIRECTORY
is not set.
Added behave test case to check if the provided master data directory
is preferred over the already set master_data_dir env variable.
it is tested by setting a wrong MASTER_DATA_DIRECTORY env variable
and when we run the utility with the correct data directory using the -d option then
the utility should execute successfully.